+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 22.10.1
 
 ### Additions and Improvements
+- Explain and improve price validation for London and local transactions during block proposal selection [#4602](https://github.com/hyperledger/besu/pull/4602)
 
 ### Bug Fixes